requirements.txt
PySide6
pytube3